# Conflicts:
#	src/main/java/com/codecrafter/WebAppManagerRoomGymBE/service/serviceimpl/NguoiDungServiceImpl.java

+++ src/main/java/com/codecrafter/WebAppManagerRoomGymBE/configsecurity/jwt/JwtToPrincipalConverter.java
@@ -14,8 +14,10 @@ public class JwtToPrincipalConverter {
         var authorityList = getClaimOrEmptyList(jwt,"au").stream()
                 .map(SimpleGrantedAuthority::new)
                 .toList();
+
+
         return UserPrincipal.builder()
-                .userId(Long.valueOf(jwt.getSubject()))
+                .userId(Integer.valueOf(jwt.getSubject()))
                 .userName(jwt.getClaim("username").asString())
                 .authorities(authorityList)
                 .build();
